interviewSchedulerUtils.js page added, for simplification
